+++ Content.Server/Shuttles/Systems/ShuttleSystem.cs
@@ -39,7 +39,6 @@ namespace Content.Server.Shuttles.Systems
             InitializeFTL();
             InitializeIFF();
             InitializeImpact();
-            InitializeShipyard();
 
             SubscribeLocalEvent<ShuttleComponent, ComponentAdd>(OnShuttleAdd);
             SubscribeLocalEvent<ShuttleComponent, ComponentStartup>(OnShuttleStartup);
@@ -63,7 +62,6 @@ namespace Content.Server.Shuttles.Systems
             CleanupEmergencyConsole();
             CleanupEmergencyShuttle();
             CleanupHyperspace();
-            CleanupShipyard();
         }
 
         public override void Shutdown()
